% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{createanClique.XCMSnExp}
\alias{createanClique.XCMSnExp}
\title{'createanClique.XCMSnExp' produces an object of class 'anClique'.}
\usage{
\method{createanClique}{XCMSnExp}(mzData)
}
\arguments{
\item{mzData}{A 'XCMSnExp' object with processed m/z data.}
}
\value{
An 'anClique' S3 object with all elements to perform clique grouping,
isotope annotation and adduct annotation.
}
\description{
\code{createanClique.XCMSnExp} creates an 'anClique' object from
'XCMSnExp' processed m/z data.
}
\examples{
require(xcms)
mzfile <- system.file("standards.mzXML", package = "cliqueMS")
rawMS <- readMSData(files = mzfile, mode = "onDisk")
cpw <- CentWaveParam(ppm = 15, peakwidth = c(5,20), snthresh = 10)
mzData <- findChromPeaks(rawMS, cpw)
ex.anClique <- createanClique(mzData)
summary(ex.anClique)
}
\seealso{
\code{\link{anClique-class}}
}
