% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{createanClique.xcmsSet}
\alias{createanClique.xcmsSet}
\title{'createanClique.xcmsSet' produces an object of class 'anClique'.}
\usage{
\method{createanClique}{xcmsSet}(mzData)
}
\arguments{
\item{mzData}{A 'xcmsSet' object with processed m/z data.}
}
\description{
\code{anClique} creates an 'anClique' object from 'xcmsSet' processed m/z data.
}
\details{
CAMERA package has to be installed to use this method.
}
\examples{
\donttest{
library(cliqueMS)
mzfile <- system.file("standards.mzXML", package = "cliqueMS")
msSet <- xcms::xcmsSet(files = mzfile, method = "centWave",
ppm = 15, peakwidth = c(5,20), snthresh = 10)
ex.anClique <- createanClique.xcmsSet(msSet)
summary(ex.anClique)
}
}
\seealso{
\code{\link{anClique-class}}
}
