% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildNetwork.R
\name{createNetwork}
\alias{createNetwork}
\title{Function to create a similarity network from processed m/z data}
\usage{
createNetwork(msSet, peaklist, filter = T, mzerror = 5e-06,
  intdiff = 1e-04, rtdiff = 1e-04)
}
\arguments{
\item{msSet}{A 'xcmsSet' object with processed m/z data}

\item{filter}{If TRUE, filter out very similar features
that have a correlation similarity > 0.99 and equal values of m/z,
retention time and intensity}

\item{mzerror}{Relative error for m/z, if relative error 
between two features is below that value that features
are considered with similar m/z value}

\item{intdiff}{Relative error for intensity, if relative
error between two features is below that value that
features are considered with similar intensity}

\item{rtdiff}{Relative error for retention time, if 
relative error between two features is below that value
that features are considered with similar retention time}
}
\value{
This function returns a list with the similarity
network and the filtered peaklist if 'filter' = TRUE. If
filter = FALSE the peaklist is returned unmodified.
}
\description{
This function creates a similarity network with nodes as features
and weighted edges as the cosine similarity between those nodes.
Edges with weights = 0 are not included in the network. Nodes
without edges are not included in the network. This network will
be used to define clique groups and find annotation within this
groups
}
\details{
Signal processing algorithms may output artifact features.
Sometimes they produce two artifact features wich are almost identical
This artifacts may lead to errors in the computation of the clique
groups, so it is recommended to set 'filter' = TRUE to drop repeated
features.
}
\examples{
netlist = createNetwork(exmsSet, exmsSet@peaks, filter = TRUE)
}
\seealso{
\code{\link{getCliques}}
}
