% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exmsSet.R
\docType{data}
\name{exmsSet}
\alias{exmsSet}
\title{Example m/z processed data}
\format{It is an 'xcmsSet' object of one sample with 276 features.}
\usage{
exmsSet
}
\description{
This dataset contains a mass sprectrometry data of metabolite standards
MS1 analyses were performed using an UHPLC system 
(1290 series, Agilent Technologies) coupled to
a 6550 ESI-QTOF MS (Agilent Technologies) operated in positive (ESI+)
electrospray ionization mode.

The metabolites in this set are the following: (-)riboflavine, 
1,2-distearoyl-sn-glycero-3-phosphocholine, biotin, cholic acid, 
deoxycholic acid, L-methionine sulfoxide, thymine and uracil
}
\keyword{datasets}
