% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{createanClique}
\alias{createanClique}
\title{'createanClique' generic function to create an object of class 'anClique'.}
\usage{
createanClique(mzData)
}
\arguments{
\item{mzData}{An object with processed m/z data. See methods for
valid class types.}
}
\value{
An 'anClique' S3 object with all elements to perform clique grouping,
isotope annotation and adduct annotation.
}
\description{
\code{createanClique} creates an 'anClique' object from processed m/z data.
}
\examples{
mzfile <- system.file("standards.mzXML", package = "cliqueMS")
library(xcms)
mzraw <- readMSData(files = mzfile, mode = "onDisk")
cpw <- CentWaveParam(ppm = 15, peakwidth = c(5,20), snthresh = 10)
mzData <- findChromPeaks(object = mzraw, param = cpw)
ex.anClique <- createanClique(mzData = mzData)
summary(ex.anClique)
}
\seealso{
\code{\link{anClique-class}}
}
