% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAnnotation.R
\name{getAnnotation}
\alias{getAnnotation}
\title{Annotate adducts and fragments}
\usage{
getAnnotation(anclique, adinfo, polarity, topmassf = 1, topmasstotal = 10,
  sizeanG = 20, ppm = 10, filter = 1e-04, emptyS = 1e-06)
}
\arguments{
\item{anclique}{Object of class 'anClique' with isotope
annotation}

\item{adinfo}{data.frame with columns 'adduct' with
adduct name, column 'log10freq' with the log10 frequency of each
adduct in the list, column 'massdiff' with the adduct mass
diff, column 'nummol' with the number of metabolite's
molecule necessary for that adduct and column 'charge' with the
charge of that adduct.}

\item{polarity}{Polarity of the adducts, choose between
'positive' or 'negative'}

\item{topmassf}{For each feature, we select a number of 
'topmassf' neutral masses that will be considered in the
final scoring of the annotation}

\item{topmasstotal}{For each group we also select a number
of 'topmasstotal' neutral masses that will be considered 
in the scoring of the annotation apart from the masses in
topmassf}

\item{sizeanG}{If a clique group has a number
of monoisotopic features bigger than 'sizeanG', we try
to divide the clique group in non-overlapping annotation groups.
Each of this subdivisions is annotated independently.}

\item{ppm}{Relative error in ppm in which we consider two or 
more features compatible with a neutral mass and two or more
adducts in 'adinfo'.}

\item{filter}{If two annotations have a relative 
mass difference smaller than 'filter' and the same features
and adducts, drop the neutral mass with less adducts}

\item{emptyS}{Score given to non annotated features}
}
\value{
An 'anClique object' with annotation columns added
to the peaklist
}
\description{
This function annotates adducts after
isotope annotation. For each clique group,
it searches for combinations of two or more features 
compatible with the same neutral mass and two or more
adducts in 'adinfo' list. For clique groups than have more
than one annotation solution, it scores all
possibilities and returns the top five solutions.
}
\details{
If clique groups have a lot of features, there are
many combinations of neutral masses and adducts. This could
lead to long running times to score the top annotations.
Parameters
'topmassf', 'topmasstotal' and 'sizeanG' are relevant in
those cases to drop the less likely neutral masses to
speed up the time of computation and still obtain
the most plausible annotation. If the clique group is small
usually no neutral masses are discarded for the scoring.
}
\examples{
library(cliqueMS)
ex.cliqueGroups <- getCliques(exmsSet, filter = TRUE)
summary(ex.cliqueGroups)
ex.isoAn <- getIsotopes(ex.cliqueGroups)
summary(ex.isoAn)
ex.adductAn <- getAnnotation(ex.isoAn, positive.adinfo, 'positive')
}
\seealso{
\code{\link{getCliques}}
\code{\link{getIsotopes}}
}
